requirements.txt
allure-pytest
pytest
python-dotenv
requests

